// frontend/lib/generated/fastapi/schemas/rewardsLeaderboardEntry.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * ShipyardHQ API
 * OpenAPI spec version: 0.1.0
 */
import type { RewardsLeaderboardEntryLastEarnedAt } from './rewardsLeaderboardEntryLastEarnedAt';
import type { RewardsLeaderboardEntryLastRedeemedAt } from './rewardsLeaderboardEntryLastRedeemedAt';
import type { RewardsLeaderboardEntryAvatarUrl } from './rewardsLeaderboardEntryAvatarUrl';

export interface RewardsLeaderboardEntry {
  userId: string;
  balance: number;
  lifetimeEarned: number;
  lifetimeSpent: number;
  lifetimeAdjusted: number;
  lifetimeRefunded: number;
  currentStreakCount: number;
  longestStreakCount: number;
  lastEarnedAt?: RewardsLeaderboardEntryLastEarnedAt;
  lastRedeemedAt?: RewardsLeaderboardEntryLastRedeemedAt;
  displayName: string;
  initials: string;
  avatarUrl?: RewardsLeaderboardEntryAvatarUrl;
  launchCount: number;
}